% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWindBarb2.R
\name{addWindBarb2}
\alias{addWindBarb2}
\title{Add wind barb to a timeseries plot}
\usage{
addWindBarb2(
  x,
  y,
  speed,
  dir,
  extraBarbLength = 0,
  barbSize = 1,
  barbLocation = "zero",
  ...
)
}
\arguments{
\item{x}{x-axis position}

\item{y}{y-axis position}

\item{speed}{wind speed in knots}

\item{dir}{wind direction in degrees clockwise from north}

\item{extraBarbLength}{add length to barbs}

\item{barbSize}{size of the barb}

\item{barbLocation}{starting location for the barb. Currently accepts 'point' or 'zero'}

\item{...}{additional arguments to be passed to \code{lines}}
}
\description{
Add a wind barb to the plot. Used internally in \link{addWindBarbs2}
}
\references{
\url{https://commons.wikimedia.org/wiki/Wind_speed}
}
\keyword{internal}
