% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWindBarbs.R
\name{addWindBarbs}
\alias{addWindBarbs}
\title{Add wind barbs to a map}
\usage{
addWindBarbs(
  x,
  y,
  speed,
  dir,
  circleSize = 1,
  circleFill = "transparent",
  lineCol = 1,
  extraBarbLength = 0,
  barbSize = 1,
  ...
)
}
\arguments{
\item{x}{vector of longitudes}

\item{y}{vector of latitudes}

\item{speed}{vector of wind speeds in knots}

\item{dir}{wind directions in degrees clockwise from north}

\item{circleSize}{size of the circle}

\item{circleFill}{circle fill color}

\item{lineCol}{line color (currently not supported)}

\item{extraBarbLength}{add length to barbs}

\item{barbSize}{size of the barb}

\item{...}{additional arguments to be passed to \code{lines}}
}
\description{
Add a multi-sided polygon to a plot.
}
\examples{
\dontrun{
library(maps)
library(RAWSmet)

setRawsDataDir("~/Data/RAWS")
cefaMeta <- cefa_loadMeta()

maps::map('state', "washington")

station1 <- cefa_load(nwsID = 451702, meta = cefaMeta) \%>\% raws_filterDate(20100623, 20100623)
station2 <- cefa_load(nwsID = 452319, meta = cefaMeta) \%>\% raws_filterDate(20100623, 20100623)

lat <- c(station1$meta$latitude, station2$meta$latitude)
lon <- c(station1$meta$longitude, station2$meta$longitude)
speed <- c(station1$data[1,]$windSpeed, station2$data[1,]$windSpeed)
dir <- c(station1$data[1,]$windDirection, station2$data[1,]$windDirection)

addWindBarbs(lon, lat, speed, dir, lwd = 2)
}

}
\references{
https://commons.wikimedia.org/wiki/Wind_speed
}
