% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWindBarbs2.R
\name{addWindBarbs2}
\alias{addWindBarbs2}
\title{Add wind barbs to a timeseries plot}
\usage{
addWindBarbs2(
  x,
  y,
  speed,
  dir,
  extraBarbLength = 0,
  barbSize = 1,
  barbColor = "blue",
  barbLocation = "zero",
  ...
)
}
\arguments{
\item{x}{vector of x-axis positions}

\item{y}{vector of y-axis positions}

\item{speed}{vector of wind speeds in knots}

\item{dir}{wind directions in degrees clockwise from north}

\item{extraBarbLength}{add length to barbs}

\item{barbSize}{size of the barb s}

\item{barbColor}{color of the barbs}

\item{barbLocation}{starting location for barbs. Currently accepts 'point' or 'zero'}

\item{...}{additional arguments to be passed to \code{arrows}}
}
\description{
Add a multi-sided polygon to a plot.
}
\examples{
\donttest{
library(RAWSmet)

data <-
  example_fw13SaddleMountain \%>\%
  raws_filterDate("2017-09-01", "2017-10-01") \%>\%
  raws_getData()

plot(data$datetime, data$windSpeed)
addWindBarbs2(data$datetime, data$windSpeed, data$windSpeed, data$windDirection)
}

}
\references{
https://commons.wikimedia.org/wiki/Wind_speed
}
