% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefa_createMeta.R
\name{cefa_createMeta}
\alias{cefa_createMeta}
\title{Obtain metadata for every station.}
\usage{
cefa_createMeta(
  metadataUrl = "https://cefa.dri.edu/raws/RAWSfw13list.xlsx",
  verbose = TRUE
)
}
\arguments{
\item{metadataUrl}{URL for station metadata.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe of station metadata.
}
\description{
Assembles individual station metadata from a CEFA webservice
into a standardized dataframe. Metdata columns include:

\itemize{
  \item{\code{deviceDeploymentID} -- unique identifier}
  \item{\code{deviceID} -- device identifier}
  \item{\code{locationID} -- location identifier}
  \item{\code{locationName} -- English language name}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{elevation} -- elevation of station in m}
  \item{\code{countryCode} -- ISO 3166-1 alpha-2}
  \item{\code{stateCode} -- ISO 3166-2 alpha-2}
  \item{\code{timezone} -- Olson time zone}
  \item{\code{nwsID} -- NWS station identifier (for FW13 data)}
  \item{\code{wrccID} -- WRCC station identifier (for WRCC data)}
  \item{\code{nessID} -- NESS station identifier (in WRCC data)}
  \item{\code{agencyName} -- responsible agency (in WRCC data)}
}
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

library(RAWSmet)

library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")

meta <- cefa_createMeta(verbose = TRUE)
dplyr::glimpse(meta)

}, silent = FALSE)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
