% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefa_load.R
\name{cefa_load}
\alias{cefa_load}
\title{Load FW13 RAWS timeseries object from a local directory}
\usage{
cefa_load(
  nwsID = NULL,
  meta = NULL,
  newDownload = NA,
  baseUrl = "https://cefa.dri.edu/raws/fw13/",
  verbose = TRUE
)
}
\arguments{
\item{nwsID}{NWS RAWS station identifier.}

\item{meta}{Tibble of FW13 station metadata.}

\item{newDownload}{Logical flag specifying whether or not to download and override existing data.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Timeseries object with 'meta' and 'data'.
}
\description{
Loads FW13 station metadata and data from the \code{rawsDataDir}.
If the data is not in this directory, this will download and save the data.
}
\note{
The `newDownload` parameter has three possible settings:
\itemize{
\item{\code{NA} -- Download data if it is not found in \code{rawsDataDir}}
\item{\code{TRUE} -- Always download data, overwriting existing data in \code{rawsDataDir}.
This is useful for updating data files with more recent data.}
\item{\code{FALSE} -- Never download data. This is useful when working with
\link{wrcc_loadMultiple} and archival data to avoid continually requesting
data for stations which have no data over a particular time period.}
}
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

library(RAWSmet)
setRawsDataDir("~/Data/RAWS/")

# For creation of metadata
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1.rda")

meta <- cefa_loadMeta()

rawsObject <- cefa_load(nwsID = 451702, meta = meta)
head(rawsObject$data)

}, silent = FALSE)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{cefa_createRawsObject}

\code{setRawsDataDir}
}
