% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefa_loadMeta.R
\name{cefa_loadMeta}
\alias{cefa_loadMeta}
\title{Load FW13 RAWS station metadata from a local directory}
\usage{
cefa_loadMeta(
  metadataUrl = "https://cefa.dri.edu/raws/RAWSfw13list.xlsx",
  newDownload = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{metadataUrl}{URL for station metadata.}

\item{newDownload}{Logical flag stating whether or not to download and override existing data.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe containing station metadata.
}
\description{
Loads all FW13 station metadata from the \code{rawsDataDir}. If
the data is not in this directory, this will download and save the data.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

# For creation of metadata
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1.rda")

library(RAWSmet)
setRawsDataDir("~/Data/RAWS")

cefa_meta <- cefa_loadMeta()

dplyr::glimpse(cefa_meta, width = 75)

}, silent = FALSE)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{cefa_createMeta}

\code{setRawsDataDir}
}
