% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefa_parseData.R
\name{cefa_parseData}
\alias{cefa_parseData}
\title{Parse RAWS FW13 characterString}
\usage{
cefa_parseData(fileString = NULL)
}
\arguments{
\item{fileString}{Character string containing RAWS FW13 data.}
}
\value{
Raw tibble of RAWS data.
}
\description{
Parse raw character data from CEFA into a tibble. The incoming \code{fileString}
can be read in directly from CEFA using \code{cefa_downloadData()} or from a
local file using \code{readr::read_file()}.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

library(RAWSmet)

tbl <-
  cefa_downloadData(nwsID = 500742) \%>\%
  cefa_parseData()

dplyr::glimpse(tbl)

}, silent = FALSE)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}

\href{https://fam.nwcg.gov/fam-web/weatherfirecd/13.htm}{FW13 Data Format}
}
\seealso{
\code{\link{cefa_downloadData}}
}
