% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cefa_meta}
\alias{example_cefa_meta}
\title{Example FW13 station metadata}
\format{
A dataframe with 66 rows and 11 columns of data.
}
\usage{
example_cefa_meta
}
\description{
The \code{example_cefa_meta} dataset provides a small subset of
FW13 station metadata containing only stations in Washington state.

This dataset was generated on 2022-01-31 by running:

\preformatted{
library(RAWSmet)
library(MazamaSpatialUtils)
library(dplyr)

setSpatialDataDir("~/Data/Spatial")

cefa_meta <- cefa_createMeta()

example_cefa_meta <- dplyr::filter(cefa_meta, stateCode == "WA")

save(example_cefa_meta, file = "data/example_cefa_meta.rda")
}
}
\keyword{datasets}
