% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cefa_Saddle_Mountain}
\alias{example_cefa_Saddle_Mountain}
\title{Example FW13 timeseries data}
\format{
A \emph{raws_timeseries object} containing \code{meta} and \code{data} dataframes.
\code{meta} has 1 row and 11 columns and \code{data} has 8744 rows and 13 columns.
}
\usage{
example_cefa_Saddle_Mountain
}
\description{
The \code{example_cefa_Saddle_Mountain} dataset provides a quickly loadable
\emph{raws_timeseries} object containing FW13 data for the station in Saddle Mountain,
Washington, (nwsID: 452701) in 2017.

This dataset was generated on 2022-01-31 by running:

\preformatted{
library(RAWSmet)

cefa_meta <- cefa_createMeta()
example_cefa_Saddle_Mountain <-
  cefa_createRawsObject(nwsID = "452701", meta = cefa_meta) %>%
  raws_filterDate(20170101, 20180101, timezone = "America/Los_Angeles")

save(example_cefa_Saddle_Mountain, file = "data/example_cefa_Saddle_Mountain.rda")
}
}
\keyword{datasets}
