% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cefaList}
\alias{example_cefaList}
\title{Example list of FW13 timeseries data}
\format{
A list of two \emph{raws_timeseries} objects each containing \code{meta} and \code{data} dataframes.
}
\usage{
example_cefaList
}
\description{
The \code{example_cefaList} dataset provides a quickly loadable list of
\emph{raws_timeseries} objects containing FW13 data for the station in Saddle Mountain, and Wellpinit
Washington, (nwsIDs: 452701, 452918) from September 2017.

This dataset was generated on 2022-01-31 by running:

\preformatted{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS/")

nwsIDs <- c("452701", "452918")
example_cefaList <-
  cefa_loadMultiple(nwsIDs = nwsIDs) %>%
  rawsList_filterDate(20170901, 20171001, timezone = "America/Los_Angeles")

save(example_cefaList, file = "data/example_cefaList.rda")
}
}
\details{
#' @title Example WRCC timeseries data
#' @format A \emph{raws_timeseries object} containing \code{meta} and \code{data} dataframes.
#' \code{meta} has 1 row and 11 columns and \code{data} has 708 rows and 9 columns.
#' @description The \code{example_wrcc_Saddle_Mountain} dataset provides a quickly loadable
#' \emph{raws_timeseries} object containing WRCC data for the station in Saddle Mountain,
#' Washington, (wrccID: waWSAD) between August 8th and September 6th 2020.
#'
#' This dataset was generated on 2020-09-06 by running:
#'
#' \preformatted{
#' library(RAWSmet)
#' setRawsDataDir("~/Data/RAWS")
#'
#' wrccMeta <- wrcc_loadMeta(stateCode = "WA")
#' example_wrcc_Saddle_Mountain <- wrcc_loadYear(wrccID = "waWSAD", meta = wrccMeta, year = 2020, password = MY_PASSWORD)
#'
#' save(example_wrcc_Saddle_Mountain, file = "data/example_wrcc_Saddle_Mountain.rda")
#' }
#'
"example_wrcc_Saddle_Mountain"
#' @title Example list of WRCC timeseries data
#' @format A list of two \emph{raws_timeseries} objects each containing \code{meta} and \code{data} dataframes.
#' @description The \code{example_wrccList} dataset provides a quickly loadable list of
#' \emph{raws_timeseries} objects containing WRCC data for the station in Saddle Mountain, and Wellpinit
#' Washington, (wrccIDs: waWSAD, waWWLP) between September 30th and October 1st 2020.
#'
#' This dataset was generated on 2020-10-1 by running:
#'
#' \preformatted{
#' library(RAWSmet)
#'
#' setRawsDataDir(~/Data/RAWS/)
#'
#' wrccIDs <- c("waWSAD", "waWWLP")
#' example_wrccList <- wrcc_loadMultiple(wrccIDs = wrccIDs, year = 2020, password = MY_PASSWORD)
#'
#' save(example_wrccList, file = "data/example_wrccList.rda")
#' }
#'
"example_wrccList"
}
\keyword{datasets}
