% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fw13_meta}
\alias{example_fw13_meta}
\title{Example FW13 station metadata}
\format{
A dataframe with 66 rows and 11 columns of data.
}
\usage{
example_fw13_meta
}
\description{
The \code{example_fw13_meta} dataset provides a small subset of
FW13 station metadata containing only stations in Washington state.

This dataset was generated on 2022-01-31 by running:

\preformatted{
library(RAWSmet)
library(MazamaSpatialUtils)
library(dplyr)

setSpatialDataDir("~/Data/Spatial")

allfw13Meta <- fw13_createMeta()

example_fw13_meta <- dplyr::filter(allfw13Meta, stateCode == "WA")

save(example_fw13_meta, file = "data/example_fw13_meta.rda")
}
}
\keyword{datasets}
