% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fw13_Saddle_Mountain}
\alias{example_fw13_Saddle_Mountain}
\title{Example FW13 timeseries data}
\format{
A \emph{raws_timeseries object} containing \code{meta} and \code{data} dataframes.
\code{meta} has 1 row and 11 columns and \code{data} has 8744 rows and 13 columns.
}
\usage{
example_fw13_Saddle_Mountain
}
\description{
The \code{example_fw13_Saddle_Mountain} dataset provides a quickly loadable
\emph{raws_timeseries} object containing FW13 data for the station in Saddle Mountain,
Washington, (nwsID: 452701) in 2017.

This dataset was generated on 2022-01-31 by running:

\preformatted{
library(RAWSmet)

fw13Meta <- fw13_createMeta()
example_fw13_Saddle_Mountain <-
  fw13_createRawsObject(nwsID = "452701", meta = fw13Meta) %>%
  raws_filterDate(20170101, 20180101, timezone = "America/Los_Angeles")

save(example_fw13_Saddle_Mountain, file = "data/example_fw13_Saddle_Mountain.rda")
}
}
\details{
#' @title Example WRCC station metadata
#' @format A dataframe with 121 rows and 11 columns of data.
#' @description The \code{example_wrcc_meta} dataset provides a small subset of
#' WRCC station metadata containing only stations in Washington state.
#'
#' This dataset was generated on 2020-09-06 by running:
#'
#' \preformatted{
#' library(RAWSmet)
#' library(MazamaSpatialUtils)
#'
#' setSpatialDataDir("~/Data/Spatial")
#' loadSpatialData("NaturalEarthAdm1")
#'
#' example_wrcc_meta <- wrcc_createMeta(stateCode = "WA")
#'
#' save(example_wrcc_meta, file = "data/example_wrcc_meta.rda")
#' }
#'
"example_wrcc_meta"
}
\keyword{datasets}
