% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fw13List}
\alias{example_fw13List}
\title{Example list of FW13 timeseries data}
\format{
A list of two \emph{raws_timeseries} objects each containing \code{meta} and \code{data} dataframes.
}
\usage{
example_fw13List
}
\description{
The \code{example_fw13List} dataset provides a quickly loadable list of
\emph{raws_timeseries} objects containing FW13 data for the station in Saddle Mountain, and Wellpinit
Washington, (nwsIDs: 452701, 452918) in 2017.

This dataset was generated on 2020-10-1 by running:

\preformatted{
library(RAWSmet)

setRawsDataDir(~/Data/RAWS/)

nwsIDs <- c("452701", "452918")
example_fw13List <- fw13_loadMultiple(nwsIDs = nwsIDs) %>% 
                      raws_filterDateList(20170101, 20180101, timezone = "America/Los_Angeles")
  
save(example_fw13List, file = "data/example_fw13List.rda")
}
}
\keyword{datasets}
