% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fw13Meta}
\alias{example_fw13Meta}
\title{Example FW13 station metadata}
\format{
A dataframe with 66 rows and 11 columns of data.
}
\usage{
example_fw13Meta
}
\description{
The \code{example_fw13Meta} dataset provides a small subset of
FW13 station metadata containing only stations in Washington state.

This dataset was generated on 2020-09-06 by running:

\preformatted{
library(RAWSmet)
library(MazamaSpatialUtils)
library(dplyr)

setSpatialDataDir("~/Data/Spatial")

allfw13Meta <- fw13_createMeta()

example_fw13Meta <- dplyr::filter(allfw13Meta, stateCode == "WA")


save(example_fw13Meta, file = "data/example_fw13Meta.rda")
}
}
\keyword{datasets}
