% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_wrcc_Saddle_Mountain}
\alias{example_wrcc_Saddle_Mountain}
\title{Example WRCC timeseries data}
\format{
A \emph{raws_timeseries object} containing \code{meta} and \code{data} dataframes.
\code{meta} has 1 row and 11 columns and \code{data} has 708 rows and 9 columns.
}
\usage{
example_wrcc_Saddle_Mountain
}
\description{
The \code{example_wrcc_Saddle_Mountain} dataset provides a quickly loadable
\emph{raws_timeseries} object containing WRCC data for the station in Saddle Mountain,
Washington, (wrccID: waWSAD) between August 8th and September 6th 2020.

This dataset was generated on 2020-09-06 by running:

\preformatted{
library(RAWSmet)
setRawsDataDir("~/Data/RAWS")

wrcc_meta <- wrcc_loadMeta(stateCode = "WA")
example_wrcc_Saddle_Mountain <- wrcc_loadYear(wrccID = "waWSAD", meta = wrcc_meta, year = 2020, password = MY_PASSWORD)

save(example_wrcc_Saddle_Mountain, file = "data/example_wrcc_Saddle_Mountain.rda")
}
}
\keyword{datasets}
