% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_wrccList}
\alias{example_wrccList}
\title{Example list of WRCC timeseries data}
\format{
A list of two \emph{raws_timeseries} objects each containing \code{meta} and \code{data} dataframes.
}
\usage{
example_wrccList
}
\description{
The \code{example_wrccList} dataset provides a quickly loadable list of
\emph{raws_timeseries} objects containing WRCC data for the station in Saddle Mountain, and Wellpinit
Washington, (wrccIDs: waWSAD, waWWLP) between September 30th and October 1st 2020.

This dataset was generated on 2020-10-1 by running:

\preformatted{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS")

wrccIDs <- c("waWSAD", "waWWLP")
example_wrccList <- wrcc_loadMultiple(wrccIDs = wrccIDs, year = 2020, password = MY_PASSWORD)

save(example_wrccList, file = "data/example_wrccList.rda")
}
}
\keyword{datasets}
