% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_wrccMeta}
\alias{example_wrccMeta}
\title{Example WRCC station metadata}
\format{
A dataframe with 121 rows and 11 columns of data.
}
\usage{
example_wrccMeta
}
\description{
The \code{example_wrccMeta} dataset provides a small subset of
WRCC station metadata containing only stations in Washington state.

This dataset was generated on 2020-09-06 by running:

\preformatted{
library(RAWSmet)
library(MazamaSpatialUtils)

setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1") 

example_wrccMeta <- wrcc_createMeta(stateCode = "WA")

save(example_wrccMeta, file = "data/example_wrccMeta.rda")
}
}
\keyword{datasets}
