% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_createMeta.R
\name{fw13_createMeta}
\alias{fw13_createMeta}
\title{Obtain metadata for every station.}
\usage{
fw13_createMeta(
  metadataUrl = "https://cefa.dri.edu/raws/RAWSfw13list.xlsx",
  verbose = TRUE
)
}
\arguments{
\item{metadataUrl}{URL for station metadata.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe of station metadata.
}
\description{
Assembles individual station metadata from a CEFA webservice 
into a standardized dataframe. Metdata columns include:

\itemize{
\item{\code{nwsID} -- NWS RAWS station identifier}
\item{\code{longitude} -- decimal degrees East}
\item{\code{latitude} -- decimal degrees North}
\item{\code{elevation} -- in meters}
\item{\code{siteName} -- human readable site name}
\item{\code{countryCode} -- ISO 3166-1 alpha-2 country code}
\item{\code{stateCode} -- ISO 3166-2 alpha-2 state code}
\item{\code{timezone} -- Olson timezone}
}
}
\examples{
\dontrun{
library(RAWSmet)

library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")

meta <- fw13_createMeta(verbose = TRUE)
dplyr::glimpse(meta)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
