% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_createMetadata.R
\name{fw13_createMetadata}
\alias{fw13_createMetadata}
\title{Obtain metadata for each station.}
\usage{
fw13_createMetadata(
  stationIDs = NULL,
  metadataUrl = "https://cefa.dri.edu/raws/RAWSfw13list.xlsx"
)
}
\arguments{
\item{stationIDs}{Vector of stationIDs to be used instead of \code{stateCode}.}

\item{metadataUrl}{Base URL for station metadata.}
}
\value{
Dataframe of station metadata.
}
\description{
Assembles individual station metadata from a CEFA webservice 
into a standardized dataframe. Metdata columns include:

\itemize{
\item{\code{countryCode} -- ISO 3166-1 alpha-2 country code}
\item{\code{stateCode} -- ISO 3166-2 alpha-2 state code}
\item{\code{stationID} -- WRCC RAWS station identifier}
\item{\code{siteName} -- human readable site name}
\item{\code{longitude} -- decimal degrees East}
\item{\code{latitude} -- decimal degrees North}
\item{\code{elevation} -- in meters}
\item{\code{timezone} -- Olson timezone}
}
}
\examples{
\dontrun{
library(RAWSmet)

meta <- fw13_createMetadata()
dplyr::glimpse(meta)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
