% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_createTimeseriesObject.R
\name{fw13_createTimeseriesObject}
\alias{fw13_createTimeseriesObject}
\title{Obtain RAWS data and create a timeseries object.}
\usage{
fw13_createTimeseriesObject(
  fw13ID = NULL,
  meta = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y-\%m-01 01:00:00 \%Z", tz =
    "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y-\%m-\%d 23:00:00 \%Z", tz =
    "UTC"),
  baseUrl = "https://cefa.dri.edu/raws/fw13/"
)
}
\arguments{
\item{fw13ID}{Station identifier found in 'meta'.}

\item{meta}{Tibble of RAWS metadata containing \code{fw13ID}.}

\item{startdate}{Desired start date (integer or character representing YYYYMMDD[HH]).}

\item{enddate}{Desired end date (integer or character representing YYYYMMDD[HH]).}

\item{baseUrl}{Base URL for data queries.}
}
\value{
Timeseries object with 'meta' and 'data'.
}
\description{
Obtains station data from a webservice and converts
it into a quality controlled, metadata enhanced "raw" tibble
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download data text}
 \item{parse data text}
 \item{standardized names}
}
}
\examples{
\dontrun{
library(RAWSmet)

meta500726 <- fw13_createTimeseriesObject(fw13ID = '500726')


}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{\link{fw13_createMetadata}}

\code{\link{fw13_createRawDataframe}}
}
