% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_load.R
\name{fw13_load}
\alias{fw13_load}
\title{Load FW13 RAWS timeseries object from a local directory}
\usage{
fw13_load(
  nwsID = NULL,
  meta = NULL,
  forceDownload = FALSE,
  baseUrl = "https://cefa.dri.edu/raws/fw13/",
  verbose = TRUE
)
}
\arguments{
\item{nwsID}{NWS RAWS station identifier.}

\item{meta}{Tibble of FW13 station metadata.}

\item{forceDownload}{Logical flag stating whether or not to download and override existing data.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Timeseries object with 'meta' and 'data'.
}
\description{
Loads FW13 station metadata and data from the \code{rawsDataDir}. 
If the data is not in this directory, this will download and save the data.
}
\examples{
\dontrun{
library(RAWSmet)
setRawsDataDir("~/Data/RAWS/")

# For creation of metadata
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1")

meta <- fw13_loadMeta()

rawsObject <- fw13_load(nwsID = 451702, meta = meta)
head(rawsObject$data)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{fw13_createTimeseriesObject}

\code{setRawsDataDir}
}
