% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_loadMeta.R
\name{fw13_loadMeta}
\alias{fw13_loadMeta}
\title{Load FW13 RAWS station metadata from a local directory}
\usage{
fw13_loadMeta(
  metadataUrl = "https://cefa.dri.edu/raws/RAWSfw13list.xlsx",
  newDownload = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{metadataUrl}{URL for station metadata.}

\item{newDownload}{Logical flag stating whether or not to download and override existing data.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe containing station metadata.
}
\description{
Loads all FW13 station metadata from the \code{rawsDataDir}. If
the data is not in this directory, this will download and save the data.
}
\examples{
\dontrun{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS/")

stationMeta <- fw13_loadMeta()

dplyr::glimpse(stationMeta)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{fw13_createMeta}

\code{setRawsDataDir}
}
