% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_loadMultiple.R
\name{fw13_loadMultiple}
\alias{fw13_loadMultiple}
\title{Load multiple FW13 RAWS timeseries objects from a local directory}
\usage{
fw13_loadMultiple(
  nwsIDs = NULL,
  meta = NULL,
  newDownload = FALSE,
  baseUrl = "https://cefa.dri.edu/raws/fw13/",
  verbose = TRUE
)
}
\arguments{
\item{nwsIDs}{Vector of NWS RAWS station identifiers.}

\item{meta}{Tibble of FW13 station metadata.}

\item{newDownload}{Logical flag stating whether or not to download and override existing data.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
List of timeseries objects each containing 'meta' and 'data'.
}
\description{
Loads FW13 station metadata and data from the \code{rawsDataDir}. If the
data is not in this directory, this will download and save the data.
}
\examples{
\donttest{
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1")

library(RAWSmet)
setRawsDataDir("~/Data/RAWS/")

stationMeta <- fw13_loadMeta()

nwsIDs <- c("021503", "500726", "020401")
stationData <- fw13_loadMultiple(nwsIDs = nwsIDs, meta = stationMeta)

dplyr::glimpse(stationData)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{fw13_createTimeseriesObject}

\code{fw13_load}

\code{setRawsDataDir}
}
