% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAWSmet.R
\name{getRawsDataDir}
\alias{getRawsDataDir}
\title{Get package data directory}
\usage{
getRawsDataDir()
}
\value{
Absolute path string.
}
\description{
Returns the package data directory where RAWS data is located.
}
\seealso{
rawsDataDir

setRawsDataDir
}
\keyword{environment}
