% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_createMetadata.R
\name{raws_createMetadata}
\alias{raws_createMetadata}
\alias{raws_getStationMetadata}
\title{Obtain metadata for each station in a state.}
\usage{
raws_createMetadata(stateCode = NULL)

raws_getStationMetadata(stationID = NULL, stateCode = NULL)
}
\arguments{
\item{stateCode}{Two character state code (will be downcased).}

\item{stationID}{Four character station ID (will be upcased)}
}
\value{
Dataframe of station metadata.
}
\description{
Obtains station metadata from a WRCC webservice and displays
it in a human-readable dataframe. The metadata includes station IDs, station
names, longitude, latitude, elevation, country codes, state codes, and 
timezones.

Passing a station ID in addition to the state code to \code{raws_getStationMetadata()}
will get a list of metadata for that specific station.
}
\examples{
\dontrun{
library(RAWSmet)

meta <- raws_createMetadata(stateCode = 'WA')
dplyr::glimpse(meta)
}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
