% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_createRawDataframe.R
\name{raws_createRawDataframe}
\alias{raws_createRawDataframe}
\title{Obtain RAWS data and parse into a tibble}
\usage{
raws_createRawDataframe(
  unitID = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
)
}
\arguments{
\item{unitID}{Station identifier (will be upcased).}

\item{startdate}{Desired start date (integer or character representing YYYYMMDD[HH]).}

\item{enddate}{Desired end date (integer or character representing YYYYMMDD[HH]).}

\item{baseUrl}{Base URL for data queries.}
}
\value{
Raw tibble of RAWS data.
}
\description{
Obtains station data from a WRCC webservice and converts
it into a quality controlled, metadata enhanced "raw" tibble
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download data text}
 \item{parse data text}
}
}
\examples{
\dontrun{
library(RAWSmet)

tbl <- raws_createRawDataframe(unitID = 'WENU')
dplyr::glimpse(tbl)
}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
\seealso{
\code{\link{raws_downloadData}}

\code{\link{raws_parseData}}
}
