% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_createTimeseriesObject.R
\name{raws_createTimeseriesObject}
\alias{raws_createTimeseriesObject}
\title{Obtain RAWS data and create a timeseries object.}
\usage{
raws_createTimeseriesObject(
  stationID = NULL,
  meta = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
)
}
\arguments{
\item{stationID}{Station identifier found in 'meta'.}

\item{meta}{Tibble of RAWS metadata containing \code{stationID}.}

\item{startdate}{Desired start date (integer or character representing YYYYMMDD[HH]).}

\item{enddate}{Desired end date (integer or character representing YYYYMMDD[HH]).}

\item{baseUrl}{Base URL for data queries.}
}
\value{
Timeseries object with 'meta' and 'data'.
}
\description{
Obtains station data from a WRCC webservice and converts
it into a quality controlled, metadata enhanced "raw" tibble
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download data text}
 \item{parse data text}
 \item{standardized names}
}
}
\examples{
\dontrun{
library(RAWSmet)

waWALD <- raws_createTimeseriesObject(stationID = 'waWALD')


}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
\seealso{
\code{\link{raws_downloadData}}

\code{\link{raws_parseData}}
}
