% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_utils.R
\name{raws_distinct}
\alias{raws_distinct}
\title{Retain only distinct data records in \code{rawsObject$data}}
\usage{
raws_distinct(rawsObject = NULL)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object}
}
\value{
A \emph{raws_timeseries} object with no duplicated data records.
}
\description{
Guarantees that records in a \emph{raws_timeseries} object
are distinct and are arranged in \code{datetime} order.
}
