% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_downloadData.R
\name{raws_downloadData}
\alias{raws_downloadData}
\title{Download RAWS data}
\usage{
raws_downloadData(
  unitID = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
)
}
\arguments{
\item{unitID}{station identifier (will be upcased)}

\item{startdate}{desired start date (integer or character representing YYYYMMDD[HH])}

\item{enddate}{desired end date (integer or character representing YYYYMMDD[HH])}

\item{baseUrl}{base URL for data queries}
}
\value{
String containing RAWS data.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a single character string containing the RAWS output.

Monitor unitIDs can be found at https://raws.dri.edu/.
}
\examples{
\dontrun{
library(RAWSmet)

fileString <- raws_downloadData(unitID = 'WENU')
print(readr::read_lines(fileString)[1:10])
}
}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
