% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_utils.R
\name{raws_extract}
\alias{raws_extract}
\alias{raws_extractData}
\alias{raws_extractMeta}
\title{Extract dataframes from \emph{raws_timeseries} objects}
\usage{
raws_extractData(rawsObject = NULL, forOpenair = FALSE)

raws_extractMeta(rawsObject)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object to extract dataframe from}

\item{forOpenair}{Logical instructions to clean data for use with openair}
}
\value{
A dataframe from the given \emph{raws_timeseries} object
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{raws_timeseries} object. These functions are designed to be useful when
manipulating data in a pipeline chain using \code{\%>\%}.
}
\examples{
\dontrun{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS")
setSpatialDataDir("~/Data/Spatial")

fw13_load(451702, year = 2005) \%>\%
  raws_extractData(forOpenair = T) \%>\%
  openair::timePlot(
    pollutant = "temperature"
  )
   
fw13_load(451702, year = 2005) \%>\%
  raws_extractMeta() \%>\%
  head()
}

}
