% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_filter.R
\name{raws_filter}
\alias{raws_filter}
\title{General purpose data filtering forraws_timeseries objects}
\usage{
raws_filter(rawsObject = NULL, ...)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object.}

\item{...}{Logical predicates defined in terms of the variables in the 
\code{rawsObject$data}.}
}
\value{
A subset of the incoming \code{raws_timeseries}.
}
\description{
A generalized data filter for \emph{raws_timeseries} objects to 
choose rows/cases where conditions are true.  Multiple conditions are 
combined with \code{&} or separated by a comma. Only rows where the condition 
evaluates to TRUE are kept.Rows where the condition evaluates to \code{NA}
are dropped.
}
\examples{
\donttest{
library(RAWSmet)

rawsObject <- example_fw13SaddleMountain

daytime <- raws_filter(rawsObject, solarRadiation > 0)
head(daytime$data)
}

}
\seealso{
\link{raws_filterDate}
}
