% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_filterDate.R
\name{raws_filterDate}
\alias{raws_filterDate}
\title{Date filtering for \emph{raws_timeseries} objects}
\usage{
raws_filterDate(
  rawsObject = NULL,
  startdate = NULL,
  enddate = NULL,
  days = NULL,
  weeks = NULL,
  timezone = NULL
)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{days}{Number of days to include in the filterDate interval.}

\item{weeks}{Number of weeks to include in the filterDate interval.}

\item{timezone}{Olson timezone used to interpret dates.}
}
\value{
A subset of the given \emph{raws_timeseries} object.
}
\description{
Subsets a \emph{raws_timeseries} object by date. This function
always filters to day-boundaries.

Dates can be anything that is understood by \code{lubridate::ymd()}
including either of the following recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}
}
\note{
The returned data will run from the beginning of \code{startdate} until
the \strong{beginning} of \code{enddate} -- \emph{i.e.} no values associated
with \code{enddate} will be returned.
}
\examples{
\dontrun{
library(RAWSmet)
setRawsDataDir("~/Data/RAWS")

rawsObject <- fw13_load(nwsID = 451702, year = 2005)

data200508 <- rawsObject \%>\% raws_filterDate(20050801, 20050901, timezone = "America/Los_Angeles")
}

}
\seealso{
\link{raws_filter}
}
