% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_filterDateList.R
\name{raws_filterDateList}
\alias{raws_filterDateList}
\title{Date filtering for a list of \emph{raws_timeseries} objects}
\usage{
raws_filterDateList(
  rawsObjects = NULL,
  startdate = NULL,
  enddate = NULL,
  days = NULL,
  weeks = NULL,
  timezone = NULL
)
}
\arguments{
\item{rawsObjects}{list of \emph{raws_timeseries} object.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{days}{Number of days to include in the filterDate interval.}

\item{weeks}{Number of weeks to include in the filterDate interval.}

\item{timezone}{Olson timezone used to interpret dates.}
}
\value{
A list of subsets of the elements of the given list of 
\emph{raws_timeseries} objects filtered by date.
}
\description{
Subsets a list of \emph{raws_timeseries} objects by date. 
This function always filters to day-boundaries.

Dates can be anything that is understood by \code{lubridate::ymd()}
including either of the following recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}
}
\note{
The returned data will run from the beginning of \code{startdate} until
the \strong{beginning} of \code{enddate} -- \emph{i.e.} no values associated
with \code{enddate} will be returned.
}
\examples{
\donttest{
library(RAWSmet)

rawsObjects <- example_fw13List

data201708 <- rawsObjects \%>\% raws_filterDateList(
                                 startdate = 20170801, 
                                 enddate = 20170901, 
                                 timezone = "America/Los_Angeles")
head(data201708)
}

}
\seealso{
\link{raws_filterDate}

\link{raws_filterList}
}
