% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_filterList.R
\name{raws_filterList}
\alias{raws_filterList}
\title{General purpose data filtering for a list of raws_timeseries objects}
\usage{
raws_filterList(rawsObjects = NULL, ...)
}
\arguments{
\item{rawsObjects}{List of \emph{raws_timeseries} object.}

\item{...}{Logical predicates defined in terms of the variables in the 
\code{rawsObject$data}.}
}
\value{
A list of subsets of the elements of the given list of 
\emph{raws_timeseries} objects.
}
\description{
A generalized data filter for a list of \emph{raws_timeseries} 
objects to choose rows/cases where conditions are true.  Multiple conditions are 
combined with \code{&} or separated by a comma. Only rows where the condition 
evaluates to TRUE are kept.Rows where the condition evaluates to \code{NA}
are dropped.
}
\examples{
\donttest{
library(RAWSmet)

rawsObjects <- example_fw13Multiple

daytime <- raws_filter(rawsObjects, solarRadiation > 0)
}

}
\seealso{
\link{raws_filterDateList}
}
