% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raws.R
\name{raws_getDataframe}
\alias{raws_getDataframe}
\alias{raws_getData}
\alias{raws_getMeta}
\title{Extract dataframes from \emph{raws_timeseries} objects}
\usage{
raws_getData(rawsObject = NULL, forOpenair = FALSE)

raws_getMeta(rawsObject = NULL)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object to extract dataframe from}

\item{forOpenair}{Logical instructions to clean data for use with openair}
}
\value{
A dataframe from the given \emph{raws_timeseries} object
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a \emph{raws_timeseries} object. These functions are designed to be useful when
manipulating data in a pipeline chain using \code{\%>\%}.
}
\examples{
\donttest{
library(RAWSmet)

data <-
  example_fw13_Saddle_Mountain \%>\%
  raws_getData(forOpenair = TRUE)

# openair::timePlot(
#  data,
#  pollutant = "temperature"
# )

example_fw13_Saddle_Mountain \%>\%
  raws_getMeta() \%>\%
  head()
}

}
