% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_utils.R
\name{raws_isRaws}
\alias{raws_isRaws}
\title{Test for correct structure for a \emph{raws_timeseries} object}
\usage{
raws_isRaws(rawsObject = NULL)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object}
}
\value{
\code{TRUE} if \code{rawsObject} has the correct structure, 
\code{FALSE} otherwise.
}
\description{
The \code{rawsObject} is checked for the 'raws_timeseries' class name
and presence of core \code{meta} and \code{data} columns.

Core \code{meta} columns include:

\itemize{
  \item{\code{nwsID} -- NWS station identifier (for FW13 data)}
  \item{\code{wrccID} -- WRCC station identifier (for WRCC data)}
  \item{\code{countryCode} -- ISO 3166-1 alpha-2}
  \item{\code{stateCode} -- ISO 3166-2 alpha-2}
  \item{\code{siteName} -- English language station name}
  \item{\code{agency} -- responsible agency (in WRCC data)}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{elevation} -- elevation of station in m}
  \item{\code{timezone} -- Olson time zone}
}

Core \code{data} columns include:

\itemize{
  \item{\code{datetime} -- measurement time (UTC)}
  \item{\code{temperature} -- temperature (C)}
  \item{\code{humidity} -- humidity (\%)}
  \item{\code{windSpeed} -- wind speed (m/s)}
  \item{\code{windDirection} -- wind direction (degrees)}
  \item{\code{maxGustSpeed} -- speed of max gust (m/s)}
  \item{\code{maxGustDirection} -- direction of max gust (degrees)}
  \item{\code{precipitation} -- precipitation (mm/h)}
  \item{\code{solarRadiation} -- solar radiation (W/m^2)}
}
}
\examples{
\donttest{
library(RAWSmet)

raws_isRaws(example_fw13SaddleMountain)
raws_isRaws(example_wrccSaddleMountain)
}

}
