% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_utils.R
\name{raws_isRawsDF}
\alias{raws_isRawsDF}
\title{Test for correct structure for a \emph{rawsDF} object}
\usage{
raws_isRawsDF(rawsDF = NULL)
}
\arguments{
\item{rawsDF}{\emph{rawsDF} object}
}
\value{
\code{TRUE} if \code{rawsDF} has the correct structure, 
\code{FALSE} otherwise.
}
\description{
The \code{rawsDF} is checked for the presence of core data columns

Core columns include:

\itemize{
  \item{\code{temperature} -- temperature (C)}
  \item{\code{humidity} -- humidity (\%)}
  \item{\code{windSpeed} -- wind speed (m/s)}
  \item{\code{windDirection} -- wind direction (degrees)}
  \item{\code{maxGustSpeed} -- speed of max gust (m/s)}
  \item{\code{maxGustDirection} -- direction of max gust (degrees)}
  \item{\code{precipitation} -- precipitation (mm/h)}
  \item{\code{solarRadiation} -- solar radiation (W/m^2)}
  \item{\code{fuelMoisture} -- fuel moisture}
  \item{\code{fuelTemperature} -- fuel temperature (C)}
  \item{\code{monitorType} -- FW13 or WRCC depending on data source}
  \item{\code{nwsID} -- NWS station identifier (for FW13 data)}
  \item{\code{wrccID} -- WRCC station identifier (for WRCC data)}
  \item{\code{siteName} -- English language station name}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{elevation} -- elevation of station in m}
}
}
\examples{
\donttest{
library(RAWSmet)

rawsDF <- example_fw13SaddleMountain \%>\% raws_toRawsDF()

raws_isRawsDF(rawsDF)
}

}
