% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_leaflet.R
\name{raws_leaflet}
\alias{raws_leaflet}
\title{Create a map of RAWS stations}
\usage{
raws_leaflet(
  meta = NULL,
  radius = 3.5,
  stroke = TRUE,
  color = "blue",
  opacity = 0.5,
  fillColor = "blue",
  fillOpacity = 0.5,
  weight = 1
)
}
\arguments{
\item{meta}{Station metadata created by \code{wrcc_createMeta()} or \code{fw13_createMeta()}.}

\item{radius}{Circle radius.}

\item{stroke}{Logical specifying whether to draw stroke along the path.}

\item{color}{Stroke color.}

\item{opacity}{Stroke opacity.}

\item{fillColor}{Fill color.}

\item{fillOpacity}{Fill opacity.}

\item{weight}{Stroke width in pixels.}
}
\value{
Leaflet map of stations.
}
\description{
Generates a map of RAWS stations given station metadata. This
station metadata can be generated by \code{wrcc_createMeta()} or \code{fw13_createMeta()}. Parameters
such as \code{radius}, \code{stroke}, \code{color}, \code{opacity}, \code{fillColor}
\code{fillOpacity}, and \code{weight} can be passed in to adjust features of
the map.
}
\examples{
\donttest{
library(RAWSmet)

wa_meta <- example_wrcc_meta
raws_leaflet(wa_meta)
}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
\seealso{
\code{\link{wrcc_createMeta}}

\code{\link{fw13_createMeta}}
}
