% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_windBarbTimeseriesPlot.R
\name{raws_windBarbTimeseriesPlot}
\alias{raws_windBarbTimeseriesPlot}
\title{Create a visualization of wind speed and direction for RAWS data.}
\usage{
raws_windBarbTimeseriesPlot(
  rawsObject = NULL,
  extraBarbLength = 0,
  barbSize = 1,
  barbColor = "blue",
  pointColor = "black",
  barbLocation = "zero",
  xlab = "Time",
  ylab = "Wind Speed",
  xlim = NA,
  ylim = NA,
  ...
)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object for which to create the plot for}

\item{extraBarbLength}{add length to barbss}

\item{barbSize}{size of the barbs}

\item{barbColor}{color of the barbs}

\item{pointColor}{color of the plotted points}

\item{barbLocation}{starting location for barbs. Currently accepts 'point' or 'zero'}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xlim}{vector containing minimum and maximum x-axis values}

\item{ylim}{vector containing minimum and maximum y-axis values}

\item{...}{additional arguments to be passed to \code{addWindBarbs2}}
}
\value{
plot of wind speed and direction
}
\description{
Creates a visualization of wind speed and direction
}
\examples{
\donttest{
library(RAWSmet)

rawsObject <- 
 example_cefa_Saddle_Mountain \%>\%
 raws_filterDate("20170901", "20170902")
  
raws_windBarbTimeseriesPlot(rawsObject, barbSize = 5, barbLocation = "point", barbColor = "blue")
}

}
