% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAWSmet.R
\docType{data}
\name{rawsDataDir}
\alias{rawsDataDir}
\title{Directory for RAWS data}
\format{
Absolute path string.
}
\description{
This package maintains an internal directory location which users can set
using \code{setRawsDataDir()}. All package functions use this directory whenever data
are created or loaded.

The default setting when the package is loaded is \code{NULL}.
}
\seealso{
getRawsDataDir

setRawsDataDir
}
\keyword{environment}
