% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawsDF_filterDate.R
\name{rawsDF_filterDate}
\alias{rawsDF_filterDate}
\title{Date filtering for \emph{rawsDF} objects}
\usage{
rawsDF_filterDate(
  rawsDF = NULL,
  startdate = NULL,
  enddate = NULL,
  days = NULL,
  weeks = NULL,
  timezone = NULL
)
}
\arguments{
\item{rawsDF}{\emph{raws_timeseries} object.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{days}{Number of days to include in the filterDate interval.}

\item{weeks}{Number of weeks to include in the filterDate interval.}

\item{timezone}{Olson timezone used to interpret dates.}
}
\value{
A subset of the given \emph{rawsDF} object.
}
\description{
Subsets a \emph{rawsDF} object by date. This function
always filters to day-boundaries.

Dates can be anything that is understood by \code{lubridate::ymd()}
including either of the following recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}
}
\note{
The returned data will run from the beginning of \code{startdate} until
the \strong{beginning} of \code{enddate} -- \emph{i.e.} no values associated
with \code{enddate} will be returned.
}
\examples{
\donttest{
library(RAWSmet)

rawsObject <- example_cefa_Saddle_Mountain
rawsDF <- rawsObject \%>\% raws_toRawsDF()

data201708 <-
  rawsDF \%>\%
  rawsDF_filterDate(20170801, 20170901, timezone = "America/Los_Angeles")

head(data201708$data)
}

}
\seealso{
\link{rawsDF_filter}
}
