% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawsList_addWindBarbs.R
\name{rawsList_addWindBarbs}
\alias{rawsList_addWindBarbs}
\title{Add wind barbs to a map for multiple RAWS stations}
\usage{
rawsList_addWindBarbs(
  rawsList = NULL,
  observationTime = NULL,
  skipMissing = FALSE,
  type = "winds",
  circleSize = 1,
  circleFill = "transparent",
  lineCol = 1,
  extraBarbLength = 0,
  barbSize = 1,
  ...
)
}
\arguments{
\item{rawsList}{a list of raws_timeseries objects}

\item{observationTime}{Time of observation in question. Only accepts POSIXct objects.}

\item{skipMissing}{Logical flag to skip stations with data missing at the requested time.}

\item{type}{Which type of wind to plot. Currently only accepts 'winds' or 'gusts'}

\item{circleSize}{size of the circle}

\item{circleFill}{circle fill color}

\item{lineCol}{line color (currently not supported)}

\item{extraBarbLength}{add length to barbs}

\item{barbSize}{size of the barb}

\item{...}{additional arguments to be passed to \code{lines}}
}
\description{
Add a multi-sided polygon to a plot.
}
\examples{
\dontrun{
library(maps)
library(RAWSmet)

setRawsDataDir("~/Data/RAWS")

maps::map('state', "washington")

# Noon on 09-01-2017
observationTime <- MazamaCoreUtils::parseDatetime(2017090112, timezone = "America/Los_Angeles")

rawsList_addWindBarbs(rawsList = example_cefaList, observationTime = observationTime)
}

}
\references{
https://commons.wikimedia.org/wiki/Wind_speed
}
