% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawsList_filter.R
\name{rawsList_filter}
\alias{rawsList_filter}
\title{General purpose data filtering for a list of raws_timeseries objects}
\usage{
rawsList_filter(rawsList = NULL, ...)
}
\arguments{
\item{rawsList}{List of \emph{raws_timeseries} objects.}

\item{...}{Logical predicates defined in terms of the variables in the
\code{rawsObject$data}.}
}
\value{
A list of subsets of the elements of the given list of
\emph{raws_timeseries} objects.
}
\description{
A generalized data filter for a list of \emph{raws_timeseries}
objects to choose rows/cases where conditions are true.  Multiple conditions are
combined with \code{&} or separated by a comma. Only rows where the condition
evaluates to TRUE are kept.Rows where the condition evaluates to \code{NA}
are dropped.
}
\examples{
\dontrun{
library(RAWSmet)

daytimeList <-
  example_fw13List \%>\%
  rawsList_filter(solarRadiation > 0)
}

}
\seealso{
\link{rawsList_filterDate}
}
