% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawsList_utils.R
\name{rawsList_isRawsList}
\alias{rawsList_isRawsList}
\title{Test for correct structure for a \emph{raws_list} object}
\usage{
rawsList_isRawsList(rawsList = NULL, verbose = FALSE)
}
\arguments{
\item{rawsList}{\emph{rawsList} object}
}
\value{
\code{TRUE} if \code{rawsObject} has the correct structure, 
\code{FALSE} otherwise.
}
\description{
The \code{rawsList} is checked for the 'list' class name and
each element of this list is checked for correct \code{raws_timeseries} object
structure.
}
\examples{
\donttest{
library(RAWSmet)

rawsList_isRawsList(example_fw13List)
rawsList_isRawsList(example_wrccList)
}

}
