% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawsList_toRawsDF.R
\name{rawsList_toRawsDF}
\alias{rawsList_toRawsDF}
\title{Convert a list of raws_timeseries objects to a single rawsDF object}
\usage{
rawsList_toRawsDF(rawsList = NULL)
}
\arguments{
\item{rawsList}{List of \emph{raws_timeseries} objects.}
}
\value{
A single tidy dataframe containing data and metadata for each
\emph{raws_timeseries} object in a \emph{raws_list}.
}
\description{
Converts a list of \emph{raws_timeseries} objects (\emph{raws_list}) 
to a single, tidy dataframe each containing all variables from each \emph{raws_timeseries}
object's \code{data} dataframe along with the following values from their \code{meta} dataframes:

\enumerate{
 \item{nwsID - the nwsID of the station}
 \item{wrccID - the wrccID of the station}
 \item{siteName - the name of the station}
 \item{longitude - longitude coordinate of the station}
 \item{latitude - latitude coordinate of the station}
 \item{timezone - the timezone the station is in}
 \item{elevation - elevation of the station}
}

This version of the RAWS data is known as a \emph{rawsDF} object.

Replicating this set of variables for every record greatly inflates the size
of the data but also makes it much more useful when working with the 
\pkg{dplyr} and \code{ggplot2} packages.

Multiple \emph{rawsDF} objects can be combined with \code{dplyr::bind_rows()} 
and used to create multi-station plots.
}
\examples{
library(RAWSmet)

rawsDF<- example_fw13List \%>\% rawsList_toRawsDF()

head(rawsDF)

}
\seealso{
\link{raws_toRawsDF}
}
