% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAWSmet.R
\docType{package}
\name{RAWSmet}
\alias{RAWSmet}
\title{Utilities for working with RAWS data.}
\description{
This package provides many functions for downloading, parsing,
and working with Remote Automatic Weather Station (RAWS) data.
}
