% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAWSmet.R
\name{removeRawsDataDir}
\alias{removeRawsDataDir}
\title{Remove package data directory}
\usage{
removeRawsDataDir()
}
\value{
Silently returns previous value of data directory.
}
\description{
Resets the package data dir to NULL. Used for internal testing.
}
\seealso{
rawsDataDir

getRawsDataDir

setRawsDataDir
}
\keyword{environment}
\keyword{internal}
