% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAWSmet.R
\name{setRawsDataDir}
\alias{setRawsDataDir}
\title{Set package data directory}
\usage{
setRawsDataDir(dataDir)
}
\arguments{
\item{dataDir}{directory where RAWS data are created}
}
\value{
Silently returns previous value of data directory.
}
\description{
Sets the package data directory where RAWS data is located.
If the directory does not exist, it will be created.
}
\seealso{
rawsDataDir

getRawsDataDir
}
\keyword{environment}
