% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesMultiplot.R
\name{timeseriesMultiplot}
\alias{timeseriesMultiplot}
\title{Faceted plot of a \emph{raws_timeseries} tibble}
\usage{
timeseriesMultiplot(
  data = NULL,
  pattern = NULL,
  parameters = NULL,
  nrow = NULL,
  ncol = NULL,
  autoRange = TRUE,
  ylim = NULL,
  style = "line"
)
}
\arguments{
\item{data}{\code{data} dataframe of a \emph{raws_timeseries} object.}

\item{pattern}{Pattern used to match groups of parameters.}

\item{parameters}{Custom vector of parameters to view.}

\item{nrow}{Number of rows in the faceted plot.}

\item{ncol}{Number of columns in the faceted plot.}

\item{autoRange}{Logical specifying whether to scale the y axis separately
for each plot or to use a common y axis.}

\item{ylim}{Vector of (lo,hi) y-axis limits.}

\item{style}{Style of plot: ("point", "line", "area")}
}
\description{
A plotting function that uses ggplot2 to display a suite of
\emph{raws_timeseries} plots all at once.
}
\note{
Specification of \code{ylim} will override the choice of
\code{autoRange}.
}
\examples{
library(RAWSmet)

data <-
  example_fw13SaddleMountain \%>\%
  raws_filterDate(20170801, 20170901) \%>\%
  raws_extractData()

timeseriesMultiplot(
  data,
  pattern = c("humidity|temperature"),
  nrow = 2
)

}
