% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windTimeseriesPlot.R
\name{windTimeseriesPlot}
\alias{windTimeseriesPlot}
\title{Create a visualization of wind and gust speeds for a given station}
\usage{
windTimeseriesPlot(rawsObject = NULL, startDate = NULL, endDate = NULL)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object for which to create the plot for}

\item{startDate}{Desired start datetime (ISO 8601).}

\item{endDate}{Desired end datetime (ISO 8601).}
}
\value{
ggplot2 plot of wind and gust speeds
}
\description{
Creates a visualization of wind and gust speeds and wind directions.
}
\examples{
library(RAWSmet)
library(grid)
library(ggplot2)

windTimeseriesPlot(example_fw13_Saddle_Mountain, 20170801, 20170802)

}
