% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createMeta.R
\name{wrcc_createMeta}
\alias{wrcc_createMeta}
\title{Obtain metadata for each station in a state.}
\usage{
wrcc_createMeta(
  stateCode = NULL,
  wrccIDs = NULL,
  baseUrl = "https://raws.dri.edu/",
  verbose = TRUE
)
}
\arguments{
\item{stateCode}{Two character state code (will be downcased).}

\item{wrccIDs}{Vector of wrccIDs to be used instead of \code{stateCode}.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe of station metadata.
}
\description{
Assembles individual station metadata from a WRCC webservice
into a standardized dataframe. Metdata columns include:

\itemize{
  \item{\code{deviceDeploymentID} -- unique identifier}
  \item{\code{deviceID} -- device identifier}
  \item{\code{locationID} -- location identifier}
  \item{\code{locationName} -- English language name}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{elevation} -- elevation of station in m}
  \item{\code{countryCode} -- ISO 3166-1 alpha-2}
  \item{\code{stateCode} -- ISO 3166-2 alpha-2}
  \item{\code{timezone} -- Olson time zone}
  \item{\code{nwsID} -- NWS station identifier}
  \item{\code{wrccID} -- WRCC station identifier (in WRCC data)}
  \item{\code{nessID} -- NESS station identifier (in WRCC data)}
  \item{\code{agencyName} -- responsible agency (in WRCC data)}
}

Because of the large number of web requests required to assemble this
metadata, it is recommended that the file be saved and reused by calling
\code{wrcc_loadMeta()}.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

library(RAWSmet)

wa_meta <- wrcc_createMeta(stateCode = 'WA', verbose = TRUE)
dplyr::glimpse(wa_meta)

}, silent = FALSE)
}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
\seealso{
\code{wrcc_loadMeta}

\code{setRawsDataDir}
}
