% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createMetadata.R
\name{wrcc_createMetadata}
\alias{wrcc_createMetadata}
\title{Obtain metadata for each station in a state.}
\usage{
wrcc_createMetadata(
  stateCode = NULL,
  wrccIDs = NULL,
  baseUrl = "https://raws.dri.edu/",
  verbose = FALSE
)
}
\arguments{
\item{stateCode}{Two character state code (will be downcased).}

\item{wrccIDs}{Vector of wrccIDs to be used instead of \code{stateCode}.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe of station metadata.
}
\description{
Assembles individual station metadata from a WRCC webservice 
into a standardized dataframe. Metdata columns include:

\itemize{
\item{\code{countryCode} -- ISO 3166-1 alpha-2 country code}
\item{\code{stateCode} -- ISO 3166-2 alpha-2 state code}
\item{\code{wrccID} -- WRCC RAWS station identifier}
\item{\code{siteName} -- human readable site name}
\item{\code{longitude} -- decimal degrees East}
\item{\code{latitude} -- decimal degrees North}
\item{\code{elevation} -- in meters}
\item{\code{timezone} -- Olson timezone}
}

Because of the large number of web requests required to assemble this 
metadata, it is recommended that the file be saved and reused.
}
\examples{
\dontrun{
library(RAWSmet)

wa_meta <- wrcc_createMetadata(stateCode = 'WA', verbose = TRUE)
dplyr::glimpse(wa_meta)
}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
