% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_downloadData.R
\name{wrcc_downloadData}
\alias{wrcc_downloadData}
\title{Download RAWS data}
\usage{
wrcc_downloadData(
  wrccID = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  password = NULL,
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
)
}
\arguments{
\item{wrccID}{Station identifier (will be upcased).}

\item{startdate}{Desired start date (integer or character representing YYYYMMDD[HH]).}

\item{enddate}{Desired end date (integer or character representing YYYYMMDD[HH]).}

\item{password}{Password required for access to archival data.}

\item{baseUrl}{Base URL for data queries.}
}
\value{
String containing RAWS data.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a single character string containing the RAWS output.

Station identifiers can be found at https://raws.dri.edu/.
}
\examples{
\dontrun{
library(RAWSmet)

fileString <- wrcc_downloadData(wrccID = 'WENU')
cat(fileString)

# Using a password for archival data
fileString <- wrcc_downloadData(
  wrccID = "WENU",
  startdate = 20200801,
  enddate = 20200901,
  password = "MY_PASSWORD"
)
}
}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
