% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_load.R
\name{wrcc_load}
\alias{wrcc_load}
\title{Load WRCC RAWS timeseries object from a local directory}
\usage{
wrcc_load(
  wrccID = NULL,
  meta = NULL,
  year = NULL,
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl",
  verbose = TRUE
)
}
\arguments{
\item{wrccID}{RAWS station identifier (will be upcased)}

\item{meta}{Tibble of FW13 station metadata.}

\item{year}{Year to access station data for.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Timeseries object with 'meta' and 'data'.
}
\description{
Loads WRCC station metadata and data from the \code{rawsDataDir}. If the
data is not in this directory, this will download and save the data.
}
\examples{
\dontrun{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS/")

stationMeta <- wrcc_createMetadata(stateCode = "WA", wrccIDs = "waWENU")
meta <- wrcc_load(wrccID = "waWENU", meta = stationMeta)

dplyr::glimpse(meta)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{wrcc_createTimeseriesObject}

\code{setRawsDataDir}
}
