% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadMeta.R
\name{wrcc_loadMeta}
\alias{wrcc_loadMeta}
\title{Load WRCC RAWS station metadata from a local directory}
\usage{
wrcc_loadMeta(
  stateCode = NULL,
  newDownload = FALSE,
  baseUrl = "https://raws.dri.edu/",
  verbose = TRUE
)
}
\arguments{
\item{stateCode}{Two character state code (will be downcased).}

\item{newDownload}{Logical flag stating whether or not to download and override existing data.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe containing station metadata.
}
\description{
Loads WRCC station metadata for a given state from the
\code{rawsDataDir}. If the data is not in this directory, this will download
and save the data.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

# For creation of metadata
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1.rda")

library(RAWSmet)
setRawsDataDir("~/Data/RAWS")

wa_meta <- wrcc_loadMeta(stateCode = "WA")

dplyr::glimpse(wa_meta, width = 80)

}, silent = FALSE)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{wrcc_createMeta}

\code{setRawsDataDir}
}
