% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadYear.R
\name{wrcc_loadYear}
\alias{wrcc_loadYear}
\title{Load WRCC RAWS timeseries object from a local directory}
\usage{
wrcc_loadYear(
  wrccID = NULL,
  meta = NULL,
  year = NULL,
  forceDownload = FALSE,
  password = NULL,
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl",
  verbose = TRUE
)
}
\arguments{
\item{wrccID}{RAWS station identifier (will be upcased)}

\item{meta}{Tibble of WRCC station metadata.}

\item{year}{Year to access station data for.}

\item{forceDownload}{Logical flag stating whether or not to download and override existing data.}

\item{password}{Password required for access to archival data.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Timeseries object with 'meta' and 'data'.
}
\description{
Loads WRCC station metadata and data from the \code{rawsDataDir}. If the
data is not in this directory, this will download and save the data.
}
\examples{
\dontrun{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS/")

stationMeta <- wrcc_loadMeta(stateCode = "WA")
rawsObject <- wrcc_loadYear(
  wrccID = "waWENU", 
  meta = stationMeta, 
  year = 2020,
  password = MY_PASSWORD
)

dplyr::glimpse(rawsObject)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{wrcc_createTimeseriesObject}

\code{setRawsDataDir}
}
