% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_parseData.R
\name{wrcc_parseData}
\alias{wrcc_parseData}
\title{Parse RAWS data string}
\usage{
wrcc_parseData(fileString = NULL)
}
\arguments{
\item{fileString}{character string containing RAWS data}
}
\value{
Dataframe of RAWS raw station data.
}
\description{
Raw character data from WRCC are parsed into a tibble.
The incoming \code{fileString} can be read in directly from WRCC using
\code{wrcc_downloadData()} or from a local file using \code{readr::read_file()}.

The type of monitor represented by this fileString is inferred from the
column names using \code{wrcc_identifyMonitorType()} and appropriate column
types are assigned. The character data are then processed, read into a tibble
and augmented in the following ways:

\enumerate{
\item{Spaces at the beginning and end of each line are removed.}
\item{All header lines beginning with ':' are removed.}
}
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

library(RAWSmet)

tbl <-
  wrcc_downloadData(wrccID = 'WSWA') \%>\%
  wrcc_parseData()

}, silent = FALSE)
}
}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
