% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cefaList}
\alias{example_cefaList}
\title{Example list of FW13 timeseries data}
\format{
A list of two \emph{raws_timeseries} objects each containing \code{meta} and \code{data} dataframes.
}
\usage{
example_cefaList
}
\description{
The \code{example_cefaList} dataset provides a quickly loadable list of
\emph{raws_timeseries} objects containing FW13 data for the station in Saddle Mountain, and Wellpinit
Washington, (nwsIDs: 452701, 452918) from September 2017.

This dataset was generated on 2022-01-31 by running:

\preformatted{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS")

nwsIDs <- c("452701", "452918")
example_cefaList <-
  cefa_loadMultiple(nwsIDs = nwsIDs) %>%
  rawsList_filterDate(20170901, 20171001, timezone = "America/Los_Angeles")

save(example_cefaList, file = "data/example_cefaList.rda")
}
}
\keyword{datasets}
