% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_createRawDataframe.R
\name{fw13_createRawDataframe}
\alias{fw13_createRawDataframe}
\title{Obtain RAWS FW13 data and parse into a tibble}
\usage{
fw13_createRawDataframe(
  nwsID = NULL,
  baseUrl = "https://cefa.dri.edu/raws/fw13/",
  verbose = FALSE
)
}
\arguments{
\item{nwsID}{Station identifier.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Raw tibble of RAWS data.
}
\description{
Obtains station data from a webservice and converts
it into a quality controlled, metadata enhanced "raw" tibble
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download data text}
 \item{parse data text}
}
}
\examples{
\dontrun{
library(RAWSmet)

tbl <- fw13_createRawDataframe(nwsID = 500742)
dplyr::glimpse(tbl)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}

\href{https://fam.nwcg.gov/fam-web/weatherfirecd/13.htm}{FW13 Data Format}
}
\seealso{
\code{\link{fw13_downloadData}}
}
