% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_createRawDataframe.R
\name{fw13_createRawDataframe}
\alias{fw13_createRawDataframe}
\title{Obtain RAWS FW13 data and parse into a tibble}
\usage{
fw13_createRawDataframe(
  stationID = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  baseUrl = "https://cefa.dri.edu/raws/fw13/"
)
}
\arguments{
\item{stationID}{Station identifier.}

\item{startdate}{Desired start date (integer or character representing YYYYMMDD[HH]).}

\item{enddate}{Desired end date (integer or character representing YYYYMMDD[HH]).}

\item{baseUrl}{Base URL for data queries.}
}
\value{
Raw tibble of RAWS data.
}
\description{
Obtains station data from a webservice and converts
it into a quality controlled, metadata enhanced "raw" tibble
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download data text}
 \item{parse data text}
}
}
\examples{
\dontrun{
library(RAWSmet)

tbl <- fw13_createRawDataframe(stationID = 500742)
dplyr::glimpse(tbl)
}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
\seealso{
\code{\link{raws_downloadData}}

\code{\link{raws_parseData}}
}
