% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_createTimeseriesObject.R
\name{fw13_createTimeseriesObject}
\alias{fw13_createTimeseriesObject}
\title{Obtain RAWS data and create a timeseries object.}
\usage{
fw13_createTimeseriesObject(
  nwsID = NULL,
  meta = NULL,
  baseUrl = "https://cefa.dri.edu/raws/fw13/",
  verbose = FALSE
)
}
\arguments{
\item{nwsID}{Station identifier found in 'meta'.}

\item{meta}{Tibble of RAWS metadata containing \code{nwsID}.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Timeseries object with 'meta' and 'data'.
}
\description{
Obtains station data from a webservice and converts
it into a quality controlled, metadata enhanced "raw" tibble
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download data text}
 \item{parse data text}
 \item{standardized names}
}


The columns of the 'data' dataframe include:

\itemize{
 \item{datetime: date and time of observation}
 \item{temperature: temperature (°C)}
 \item{humidity: average relative humidity (\%)}
 \item{windSpeed: average wind speed (m/s)}
 \item{windDirection: average wind direction (°)}
 \item{maxGustSpeed: max gust speed (m/s)}
 \item{maxGustDirection: max gust direction (°)}
 \item{precipitation: precipitation}
 \item{solarRadiation: solar radiation (W/m^2)}
 \item{fuelMoisture: 10-hour time lag fuel moisture}
 \item{fuelTemperature: NA for FW13 data}
}
}
\examples{
\dontrun{
library(RAWSmet)

library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")

nws_500726 <- fw13_createTimeseriesObject(nwsID = 500726)

}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}

\href{https://fam.nwcg.gov/fam-web/weatherfirecd/13.htm}{FW13 Data Format}
}
\seealso{
\code{\link{fw13_createMetadata}}

\code{\link{fw13_createRawDataframe}}
}
