% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_downloadData.R
\name{fw13_downloadData}
\alias{fw13_downloadData}
\title{Download RAWS FW13 data}
\usage{
fw13_downloadData(
  stationID = NULL,
  baseUrl = "https://cefa.dri.edu/raws/fw13/"
)
}
\arguments{
\item{stationID}{Station identifier.}

\item{baseUrl}{Base URL for data queries.}
}
\value{
String containing RAWS data.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a single character string containing the RAWS output.

Station identifiers can be found at https://cefa.dri.edu/raws/RAWSfw13list.xlsx.
}
\examples{
\dontrun{
library(RAWSmet)

fileString <- fw13_downloadData(stationID = 500742)
print(readr::read_lines(fileString)[1:10])
}
}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
