% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw13_downloadData.R
\name{fw13_downloadData}
\alias{fw13_downloadData}
\title{Download RAWS FW13 data}
\usage{
fw13_downloadData(nwsID = NULL, baseUrl = "https://cefa.dri.edu/raws/fw13/")
}
\arguments{
\item{nwsID}{NWS RAWS station identifier.}

\item{baseUrl}{Base URL for data queries.}
}
\value{
String containing RAWS data.
}
\description{
Request data for a particular station. Data are returned as a
single character string containing the RAWS output.

Station identifiers can be found at https://cefa.dri.edu/raws/RAWSfw13list.xlsx
and can be downloaded with \code{\link{fw13_createMeta}}.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

library(RAWSmet)

fileString <- fw13_downloadData(nwsID = 500742)
print(readr::read_lines(fileString)[1:10])

}, silent = FALSE)
}
}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
