% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_downloadData.R
\name{raws_downloadData}
\alias{raws_downloadData}
\title{Download RAWS data}
\usage{
raws_downloadData(
  stationID = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
)
}
\arguments{
\item{stationID}{Station identifier (will be upcased).}

\item{startdate}{Desired start date (integer or character representing YYYYMMDD[HH]).}

\item{enddate}{Desired end date (integer or character representing YYYYMMDD[HH]).}

\item{baseUrl}{Base URL for data queries.}
}
\value{
String containing RAWS data.
}
\description{
Request data from a particular station for the desired time period.
Data are returned as a single character string containing the RAWS output.

Station identifiers can be found at https://raws.dri.edu/.
}
\examples{
\dontrun{
library(RAWSmet)

fileString <- raws_downloadData(stationID = 'waWENU')
print(readr::read_lines(fileString)[1:10])
}
}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
