% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws_filterDate.R
\name{raws_filterDate}
\alias{raws_filterDate}
\title{Date filtering for \emph{raws_timeseries} objects}
\usage{
raws_filterDate(
  rawsObject = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  unit = "sec",
  ceilingStart = FALSE,
  ceilingEnd = FALSE
)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates.}

\item{unit}{Units used to determine time at end-of-day.}

\item{ceilingStart}{Logical instruction to apply
\code{\link[lubridate]{ceiling_date}} to the \code{startdate} rather than
\code{\link[lubridate]{floor_date}}}

\item{ceilingEnd}{Logical instruction to apply
\code{\link[lubridate]{ceiling_date}} to the \code{enddate} rather than
\code{\link[lubridate]{floor_date}}}
}
\value{
A subset of the given \emph{raws_timeseries} object.
}
\description{
Subsets a \emph{raws_timeseries} object by date. This function
always filters to day-boundaries.

Dates can be anything that is understood by \code{lubridate::ymd()}
including either of the following recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}
}
\note{
The returned data will run from the beginning of \code{startdate} until
the \strong{beginning} of \code{enddate} -- \emph{i.e.} no values associated
with \code{enddate} will be returned. The exception being when
\code{enddate} is less than 24 hours after \code{startdate}. In that case, a
single day is returned.
}
\examples{
\donttest{
library(RAWSmet)

data201708 <-
  example_fw13_Saddle_Mountain \%>\%
  raws_filterDate(20170801, 20170901, timezone = "America/Los_Angeles")

head(data201708$data)
}

}
\seealso{
\link{raws_filter}
}
