% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raws.R
\name{raws_isEmpty}
\alias{raws_isEmpty}
\title{Test for an empty \emph{raws_timeseries} object}
\usage{
raws_isEmpty(rawsObject = NULL)
}
\arguments{
\item{rawsObject}{\emph{raws_timeseries} object}
}
\value{
\code{TRUE} if no data exists in the \code{raws_timeseries} object.
\code{FALSE} otherwise
}
\description{
Conveniently tests if a given \emph{raws_timeseries} object
is empty.
}
\examples{
\donttest{
library(RAWSmet)

raws_isEmpty(example_cefa_Saddle_Mountain)
raws_isEmpty(example_wrcc_Saddle_Mountain)
}

}
