% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createRawDataframe.R
\name{wrcc_createRawDataframe}
\alias{wrcc_createRawDataframe}
\title{Obtain RAWS data and parse into a tibble}
\usage{
wrcc_createRawDataframe(
  wrccID = NULL,
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  password = NULL,
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
)
}
\arguments{
\item{wrccID}{Station identifier (will be upcased).}

\item{startdate}{Desired start date (integer or character representing YYYYMMDD[HH]).}

\item{enddate}{Desired end date (integer or character representing YYYYMMDD[HH]).}

\item{password}{Password required for access to archival data.}

\item{baseUrl}{Base URL for data queries.}
}
\value{
Raw tibble of RAWS data.
}
\description{
Obtains station data from a WRCC webservice and converts
it into a quality controlled, metadata enhanced "raw" tibble
ready for use with all \code{raw_~} functions.

Steps involved include:

\enumerate{
 \item{download data text}
 \item{parse data text}
}
}
\examples{
\dontrun{
library(RAWSmet)

tbl <- wrcc_createRawDataframe(wrccID = 'waWENU')
dplyr::glimpse(tbl)
}

}
\references{
\href{https://raws.dri.edu/}{RAWS USA Climate Archive}
}
\seealso{
\code{\link{wrcc_downloadData}}

\code{\link{wrcc_parseData}}
}
