% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadMeta.R
\name{wrcc_loadMeta}
\alias{wrcc_loadMeta}
\title{Load WRCC RAWS station metadata from a local directory}
\usage{
wrcc_loadMeta(
  stateCode = NULL,
  baseUrl = "https://raws.dri.edu/",
  verbose = TRUE
)
}
\arguments{
\item{stateCode}{Two character state code (will be downcased).}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Dataframe containing station metadata.
}
\description{
Loads WRCC station metadata for a given state from the \code{rawsDataDir}. If the
data is not in this directory, this will download and save the data.
}
\examples{
\dontrun{
library(RAWSmet)

setRawsDataDir("~/Data/RAWS/")

stationMeta <- wrcc_loadMeta(stateCode = "WA")

dplyr::glimpse(stationMeta)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{wrcc_createMetadata}

\code{setRawsDataDir}
}
