% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_loadMultiple.R
\name{wrcc_loadMultiple}
\alias{wrcc_loadMultiple}
\title{Load multiple WRCC RAWS timeseries objects from a local directory}
\usage{
wrcc_loadMultiple(
  wrccIDs = NULL,
  meta = NULL,
  year = NULL,
  forceDownload = FALSE,
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl",
  verbose = TRUE
)
}
\arguments{
\item{wrccIDs}{Vector of RAWS station identifiers (will be upcased).}

\item{meta}{Tibble of WRCC station metadata.}

\item{year}{Year to access station data for.}

\item{forceDownload}{Logical flag stating whether or not to download and override existing data.}

\item{baseUrl}{Base URL for data queries.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
List of timeseries objects each containing 'meta' and 'data'.
}
\description{
Loads WRCC station metadata and data from the \code{rawsDataDir}. If the
data is not in this directory, this will download and save the data.
}
\examples{
\donttest{
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1")

library(RAWSmet)
setRawsDataDir("~/Data/RAWS/")

stationMeta <- wrcc_loadMeta(stateCode = "WA")

wrccIDs <- c("waWENU", "waWASH", "waWALD")
stationData <- wrcc_loadMultiple(wrccIDs = wrccIDs, meta = stationMeta, year = 2020)

dplyr::glimpse(stationData)
}

}
\references{
\href{https://cefa.dri.edu/raws/}{Program for Climate, Ecosystem and Fire Applications}
}
\seealso{
\code{wrcc_createTimeseriesObject}

\code{wrcc_load}

\code{setRawsDataDir}
}
